Move Tracing class next to other tracing classes to keep things consistent
