+++ phpstan.neon
@@ -8,7 +8,7 @@ parameters:
         - tests
     excludes_analyse:
         - tests/End2End/App
-        - tests/EventListener/Tracing/RequestListenerTest.php
+        - tests/Tracing/Symfony/RequestEventsListenerTest.php
     dynamicConstantNames:
         - Symfony\Component\HttpKernel\Kernel::VERSION
         - Doctrine\DBAL\Version::VERSION
